Ensure that the password is hashed on db creation.
This code is taken from the openldap_database type. The rootpw
property has logic to hash the password, if it isn't already hashed,
and ensure that the database matches. However, when creating the
database resource for the first time that code isn't executed and
instead create is called on the provider. By duplicating the logic
here we can ensure that the password is correctly hashed when the
database is created.

This fixes #101 where the password is added unhashed during the
first run but fixed on the next Puppet run when the logic in the
type can take effect.

Minor niggle: duplicating code isn't ideal.