Workaround CI failure
For some reason, when an openldap::server::overlay is in place in GitHub
Action CI, the slapd service fail to start.

I was not able to reproduce this issue on Docker nor Vagrant on my
machine, and I suspect the issue is tight to the execution envrionment
probided by GitLab Actions.

As a workaround, add a test that remove overlays.  This is enough to
unbreak the CI.