Switch to regular OpenLDAP packages on FreeBSD
FreeBSD opeldap packages useds to be compiled without SASL support,
needed for OLC configuration to work, so the module default to these
versions of the package with SASL support.

Recently [1], the ports where updated so that the base package have SASL
support, and non-SASL OpenLDAP is not available anymore on FreeBSD.  As
a consequence, the openldap-sasl-client and openldap-sasl-server do not
exist anymore.

Update the names of the default packages used on FreeBSD to match what
is actualy the default.

References:
  1. https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=257374