Get rid of dummy number in openldap::server::access_wrapper::acl
The role of these rumbers was to allow multiple ACL with the same `to`
to not overwrite each other when placed into the same hash.  They have
no further purpose and do not really avoid the issue since one can reuse
the number.

Rework the expected data structure to allow constructs that do not allow
such misconfiguration.