Allow to skip management of policy-rc.d
Unfortunately, we can't count on a better system than `policy-rc.d(8)`
on Debian for preventing the service from starting on package
installation.  Some systems administrators might have system-wide
configuration to never start any service, and they will not want to have
this file managed in an uncontrolled way by the module.

We assume that these administrators already have a `file` resource to
manage this script, and that the default configuration will cause
conflicts and the catalog will not be built.  By disabling this script
management in the module, administrators will revert to their previous
behavior, after having a chance to adjust their own script if needed.