Allow to have Sensitive[String[1]] for rootdn and rootpw.
Currently only rootpw can be assigned as Sensitive, but in some cases one might want to keep the username (DN) out of the PuppetDB and the logs.

In our case, we use a secret lookup function that automatically wraps the data into a Sensitive type, and we store the DN and password together.